--- README.md
@@ -1,3 +1,0 @@
-Thi is a try to make a game in C, similar to space invaders, that work on comand shell only in Windows.
-I working on a cross-platform version of the code.
-
