--- dati.dat
@@ -1,1 +1,0 @@
-50 15 2 0 no_record
